Add notebook for manipulating district geometry files
The notebook reads the district geometry data from the city council and cleans it up and changes the reference point to the centre of the city (in the case of Madrid, Puerta del Sol), and the units to km.

It then saves the changes as a new geometry file, which will be used in the city simulation.