Eliminated geopandas dependency in city simulator
The dependency on geopandas was preventing the simulator to work properly on Windows systems due to a bug in the library. It has been substituted by Fiona, and the overall management of geometries in the simulator has been streamlined.

The geopandas dependency remains in the project overall, since it is still being used in a few notebooks.