feat: Add identifier to ambulance and emergency objects
This identifier is assigned incrementally as emergencies appear or ambulances are launched in time, per severity level. Identifiers are unique only within the same severity level.

The identifier assigned to an ambulance is the same as that of the emergency it is acting upon, or in the case of an ambulance transference between hospitals, it is the number of transferences up to that point in history.