feat: Change reward to be distributed through time
Instead of reward being adjudicated all at once when the ambulance reaches the destination, it is now adjudicated bit by bit at every step in which the emergency has not been attended yet. This sidesteps the possible corner case in which the agent learns not to launch any ambulances in order to avoid any costs being applied, and in general will likely make learning easier due to less sparse rewards.